+++ CHANGELOG.md
@@ -1,7 +1,29 @@
+## 0.5.1
+
+- BCDice Ver2.02.73 にアップデート
+- Sinatra 2.0 に更新
+- 依存するその他のGemのバージョンを更新
+
+## 0.5.0
+
+- BCDice Ver2.02.72 にアップデート
+- 処理の一部を最適化 (thanks @ochaochaocha3)
+
+
+## 0.4.0
+
+- JSONPに対応
+- 失敗時に妥当なHTTPステータスコードを示すように
+- 失敗時のレスポンスに `reason` を追加
+- DiceBotのsysteminfoで標準ダイスボットのヘルプを返すように
+- Herokuに直接デプロイできるように設定を記述
+- [for developer] テストコードを作成
+
+
 ## 0.3.0
 
-- `/v1/systeminfo` (thanks @kumakaba)
-- `/v1/diceroll`でダイスごとの結果を返すように (thanks @kumakaba)
+- BCDice Ver2.02.70 にアップデート
+- extratablesに対応 (thanks @kumakaba)
 
 
 ## 0.2.0